Fold `nativeFromFabricValue` round-trip coverage into `fabric-value.test.ts`; drop dispatch test.
`fabric-value-dispatch.test.ts` tested encode+decode round-trips through the
data-model flag, which `fabric-value.test.ts` didn't cover at all (it has no
`nativeFromFabricValue` tests). Preserve the non-redundant modern bits as a new
`nativeFromFabricValue` describe: round-trip of primitives, `undefined`,
`bigint`, objects, arrays, and a nested structure, plus `FabricError`->`Error`
unwrapping. The legacy-path and config-lifecycle (flag machinery) blocks are
dropped, and the dispatch test file is deleted.
